Extract about and skills sections to own components
